Reduce the Onslaught of Uniques
With a new option, of course. User may decide what percentage of
uniques to face in this game, from 10% all the way up to 100% (the
default). The game chooses random quest monsters and bounty monsters
before reducing uniques. Also, dungeon guardians and FIXED_UNIQUE
monsters are excluded from the reduction. Thus, in practice, you
will face a few more monsters than your percentage indicates. Another
way to think of this is that you specify the percentage of "qualifying"
monsters to face. You'll still need to kill Oberon and the Serpent :)

This is a first pass implementation, so there may still be bugs.