Quickband: A new birth option for much quicker/easier play
I've long wanted to add this, but wasn't sure how best to do it.
Now that we have reduce_uniques, I think this will make a plausible
substitute for the Munchkin personality.

Try the following:
  no_wilderness  = ON
  reduce_uniques = 20%
  quickband      = ON
  [random arts   = 70%]

Of course, this probably needs a lot of tweaking, but the goal is
to be able to win the game in a single day, perhaps a single sitting.