sdl: experimenting with fonts. cleanup colors
I wanted to experiment with using native system fonts. This is possible,
but the results aren't very satifying. For one thing, there is no anti-
aliasing. For another, often times the letters are ridiculously off
center, something which fails to happen when using the same font in the
curses port (which uses Konsole's wonderful font handling).

In order to make this experiment, I had to raid Vanilla's main-sdl.c for
a more modern SDL port. This is not easy!