oook: quest reward lost when pack is full
Need to delay pack_overflow() when p_ptr->leaving ... Receiving
a quest reward forces the wilderness to regenerate once you leave
the building. If your pack is full, the reward goes to the overflow
stack, which was supposed to get spilled on town floor. Unfortunately,
pack_overflow() was immediately followed by regenerating the town,
so any objects on the ground were lost. Fixed (and tested this time!)