Sniper: Recode for more informative UI
Damage calcs are unchanged, for the most part. Except that VULN_COLD
and VULN_FIRE monsters are now hurt more by SP_COLD and SP_FIRE. Also,
there is a slight benefit if you are using ammo that has the same
effect. For example, OF_BRAND_FIRE with SP_FIRE or OF_SLAY_EVIL with
SP_HOLYNESS.

The main changes, however, focus on making the Sniper's game mechanics
more transparent to the player. For example, displaying the correct
accuracy for the current level of focus; showing the slay multiplier
in the spell list; and - most important of all - making spell browsing
redirect to display the shooter info with the correct damage listed.

A recent complaint on oook actually had a 2k shooting option with 88%
accuracy, yet the player missed this and won using magical devices
instead (despite that probably being about 10x more difficult to do).
If played correctly, the sniper's end battle is trivial, but getting
there may take some work.