Inventory: Display Slots (cont)
Here is some work wrt identify and The Mogaminator. Auto usage of
scrolls of identify now receives a delayed treatment, as does mass
identification (and even normal id) ... at least for INV_PACK and
INV_QUIVER. This will need a bit of playtesting, though.

Some "You get ..." messages have been removed. I'm not a fan of overly
spamming the user with messages, so, if we are going to display "You
have ..." messages, then I'd rather just leave it at that. For example:
  Really buy 12 potions of foo for 12gp?
  You bought 12 potions of foo for 12gp.
  You have 23 potions of foo (e).
The second message is noise, IMO.