Travel: Improve Route Selection
Previously, travel_step() would inappropriately choose erratic
diagonal paths when traveling. On the surface, this often induces
an unnecessary wilderness scroll, potentially losing the target
in the process. Also, the diagonal paths look like the player is
a "singing, happy drunk"! I changed the algorithm to *not* favor
the previous travel.dir, and to choose horizontal and vertical
directions in preference to diagonals when discriminating equal
cost alternatives.