Snipers: OP vs Serpent
Paradoxically, it is not so much the insane damage, as it is the
fact that stunning barely effects fail rates on ?Teleport. At some
point in the past, there was a minimum fail on devices when stunned,
but it must have changed in Chengband. Restore 15% min fail rates
(25% if heavily stunned). With this change, SP_FINAL can actually
go up in damage (from 2000+ to 2500+ with reasonable gear). SP_FINAL
should be the best multiplier. While SP_HOLYNESS is normally 3.8x max,
it is actually 6.6x max against light vulnerable evil monsters. Thus,
SP_FINAL should be at least that. I made it 7.0x.