display_percentages in status bars rather than [****---]
This came up when playtesting the Sniper. With this option on,
you can use Probing and the more accurate change in monster health
percent to get a fairly accurate in-game idea of how deadly your
shots are. But all classes will probably enjoy this change as
the old status bar was often misleading due to rounding issues.