oook bug: shapeshifting and the quiver
Shifting to a form that lacks a quiver slot (e.g. hound) removes
the quiver from your equipment, but doesn't handle ammo. In fact,
it is illegal to remove a loaded quiver, so the bug is that shape-
shifting doesn't handle this. Now it does, marking ammo as <<Worn>>
so it gets re-equipped in the quiver once you shift back to an
appropriate form.