oook report: un-necessary pack overflow when wielding
This happens if you wield with a full pack. Wielding first removes
the current object in the requested slot, but your pack is full, so
it registers it in _overflow. Then, it moves the chosen object to
your equipment, freeing up a slot after all. Due to memory ownership
issues, its simplest to just have pack_overflow detect this case.