oook request: display object list from do_cmd_look
I didn't even know this existed ... You needed to turn on easy_floor
to get this behavior in the old code. I tried to reverse engineer
what the old code did ... List navigation is clumsy since we are in
the middle of what is actually the targetting code. Arrows, + and -
keys ... these all mean something else in the broader context. So we
can't have something as nice as obj_prompt. Still, you can scroll down
with ENTER and scroll up with '\' (which is right above your ENTER key).