Quiver is more discriminating in its affections
Previously promiscuous, the quiver now only likes matching ammo that
can combine with an existing pile, or that is inscribed with '=g'.
This last restriction is for artifact ammo as well as for the common
case of completely discharging an entire pile. Players should inscribe
'=g' on the ammo they are using (together with @fX directives too!).