Linux Buildsystem: Remove memory leak from autoconf program
In particular, AM_PATH_SDL creates a program that mallocs a format
string, but never frees it. When returning 0, ASAN detects and reports
a memory leak, and changes $? to 1 causing configure to think sdl
is out of date.

In general, don't sanitize your configure. Use $SANITIZE_FLAGS instead.
But in case of accidents, I'd like to save you a really bad headache.