oook request: problems with duplicate inscriptions
There are many ways to handle this, the easiest being to not use
duplicate inscriptions. However, with autoinscriptions from the
Mogaminator, this would require extra work. The quickest thing to
do right now is to add a command '@' to toggle the usage of inscription
label overrides on and off. That way, if you are ever in a situation
where an object choice is inaccessible, simply tap '@' and then use
the normal inventory letter to choose that object. This override
resets on each obj_prompt, so your next prompt will go back to using
the normal inscription overrides.

This complaint also happens to me quite a bit. For example, I do
'@rb' on scrolls of phase door. Often, this then blocks another scroll
in the b slot. Usually, if I inscibe a class of objects, though, I
end up inscribing them all and only get burned when finding new flavors.