mon_display: always show r_info spell_freq if easy_lore
By request, even if it is wrong (and it tends to be). I tried to explain
my reason for not using spell_freq in an overly verbose comment. There
is not really room to display both frequencies. I suppose we could not
show the "(11 of 22 moves)" if easy_lore is turned on. The complaint is
that they don't care about that anyway, so that might be best.