Rage-Mage: Minor Changes
AC squishiness is now non-linear, so less of an early game beat down.
Mana Clash damage recoded since, for example, The Serpent is not a 1 in
anything caster. It is a 40% caster (100/40 => 1 in 2, which is wrong).
Finally, Berserk now helps with the Anti-magic Ray, but only by
requiring monsters to make an extra save. The duration is till very
short.