Rage-Mage: Worse Skills, but Less Berserk Biff
This class is not a device class. They should not avoid using Berserk so
they can use powerful devices. To this end, I decreased their class
based device skills to become more warrior-like, but also decreased the
amount of the Berserk skill penalty. There is some drawback to using
Berserk, but also much advantage, and this talent should be used as much
as possible with this class.