Devices: Upper tier devices are not for everybody.
Ah, devices. Apparently these make winning much easier? Well, there
needs to be a wall in the difficulty curve. I mean, if non-device
classes like the Rage-Mage can use -Rockets with impunity, why play a
Mage? This is silliness.

Lots of changes in this one, mostly "bad". But, to compensate, _Healing is
now an L55 effect. Believe me, this does make a huge difference! Oh, and
devices of simplicity are less silly. As usual, consult the spreadsheet
in design/devices.ods when evaluating things.