Missing Death Dumps
It is simpler to just prompt the user to make sure, no? You now have
to be really mashing the keyboard in rage at your unfair death to miss
getting a dump (requires you (1) not to read the message, (2) to press
ESC, then (3) not to read the warning about missing a dump and to mash
the y key in total anger).