Spoilers: Monster Damage by Resistance
I moved this from spoilers.c (created csv file) to the on-line wizard
spoilers screen (^A"d). Now, we create a document which the user may
view online or save to html, etc. In addition, the display reflects
the current character status. Damage amounts are adjusted by current
resistances and colored based on current hitpoints.

This is information I use for game balance and is not intended for
normal play, but some people like spoilers so I though I'd share.