Revert 7ed2f77: auto_id does identify_pack in process_world
Originally, I accepted this change since I thought it was to address
problems with amnesia. But now, I think it was more to make the Stone
of Lore just as good when used as a swap. This is *bad*!

auto_id is a very strong ability. The Stone of Lore got it (cf this
object in Hengband to see how amazing a boost this is) under the
assumption that the user would wear the item for the benefit, thereby
forgoing other more useful light sources during the early game (such as
Phial of Galadriel replacements or early reforges ... These usually have
stat boosts and/or a resistance or two and are a competitive choice with
The Stone of Lore). Players should not use this item as a swap.

I can see how the logic went when looking at some of the Poschengband
fork repositories:
[1] Make Stone of Lore convenient to swap.
[2] Now, why waste keystrokes. Make Stone of Lore work from pack.
[3] Now, why waste a slot, make every player "evolve into a Stone of
Lore" at CL25.
[4] Now, feel guilty and remove an inventory slot.

I think easy_id is generous as is without these sorts of additions. If
you want auto_id, you need to wear something, or play one of the (many)
races/classes that let you choose this as a bonus power. Granting it for
free is a *bad* idea.

auto_id now protects the player from object related amnesia effects, but
no longer applies during process_world.