The Sand Pits: Change Reward to Match Quest
I don't like having another light source artifact reward, especially
The Stone of Lore (It promotes delaying the quest until you find the
original, thereby getting the replacement art). Many of the quests
choose rewards that match the quest. Since you need Chaos resistance to
withstand the ELDRITCH_HORROR, let the reward grant that: a Blade of
Chaos. (This may be too generous for QL38, but Bloodletters are ML34 and
drop blades of chaos often enough.)