Quests: Keep the player guessing
These types of changes make me nervous, but it seems to be working.
Before, if the player observed a downshaft on a level, they could safely
concluded that the next level was not a quest level. And if a level
lacked downshafts, then they could be certain the next level was indeed
a quest level. Thus, the aim of randomizing quest levels was subverted.
Instead, allow downshafts to spawn as normal, but change them to work
just like normal stairs when the next level is a quest level.

This needs more testing to make sure I got all the required bits of
code.