Rand Arts: Fix Power Limits
Now that I have decent slot-based scoring to normalize artifacts based
on slot type, I reverted the previous commit to tone down the power of
end game rand-arts. My experiences were with overpowered *auxillary*
armor slots (gloves, boots, cloaks and helms), not with body armor or
melee weapons.

In the process, I uncovered a major bug when scoring an unidentified
artifact. Even with the COST_REAL option, the code was assuming the
player's current level of knowledge and applying the "not an artifact"
scoring biff (3/4 real score). Thus, rand arts have been quite a bit
more powerful then the previously set limits (4/3 or 33% overpowered).
Weak slots should only be 50% or less of full powered slots, so this
overshot was by as much as 300%, which is rather noticeable!