Linux Buildsystem: Problems with gcc 6
Grabbed posted change from oook. However, this still fails to work for
me with gcc 6.2 as ld does not support a --push-state option. Somehow, I
probably need to find an updated ld (gcc-6 definitely spawns ld when
linking), but I have no idea how or where to look. So, I'm not sure if
this fix actually works for the OP ...