Artifacts: Replacements and Scoring
Add FIXED_ACT to preserve activations of artifacts during replacement.
For example, the two guns keep their "fire" activations for flavor.

Downplay the value of immunities to the value of 3 high resists or so.
This will help keep replacements from going nutty.

Downplay the value of several good activations, keeping them about the
same value as an immunity (e.g. Heal 1000 on Soulkeeper or Restore Mana
on Daeron).