Archery: Brand and Crimson
To encourage archery, I'm trying to tone down the alternatives. One such
is the use of the bow slot as a speed enhancement slot. If I remove
speed from Brand, it is too much like Hellfire to be worth keeping
around, so I removed it (QUESTITEM to preserve upgrades). Crimson can
safely lose the speed boost and return the XTRA_POWER and XTRA_H_RES
flags.