Sekhmet: Attempt a Rescue
Again, ridiculously OP. I nerfed down to about the strength of Gothmog,
though perhaps a bit nastier with the cuts and stuns (Gothmog lists 216
melee while Sekhmet reduced from 207 to 163, but this does not account
for the extra cut damage which more than makes up the difference. Both
uniques are about 8 to 10 !Healing and 2 !*Healing* for an OP player.)

BTW, Gothmog is supposed to be as tough as they come, so, again, nothing
stronger please! The healing amounts listed are about 10% of the game
totals (unless you can use devices), so we can't have too many uniques
of this caliber.