Typhoeus: Tone Down
More ridiculosity ... Compare Surtur and Ymir, both who should at least
be in the same league. Although, Typhoeus should be the strongest of the
three. His storm breath is a bit un-nerving, so he has no business being
a 1 in 2 caster (Ymir is 1 in 3 while Surtur just 1 in 5). I made him 1
in 3.