Shesha: Tone down melee damage
Again, blind and confuse hits were too strong ... There's no AC damage
reduction for those effects. Shesha was considerably stronger than
Cyaegha in melee as a result. With 50% more hp than the latter and a
BR_DISI summon naga combo, it was too much. They are both DL80 and I
consider Cyaegha to be a high water mark for monster power at that
depth. I left hp as is, but they could probably come down. Most nearby
uniques are in the 6 to 7k range.