New Monsters: Adjust HP
A steady glance at mon-damage.html (^A"d) really helps identify
outliers. Unless there is a strong reason for 20% to 50% more hp then
comparably depthed monsters, things got nerfed (I even repaired Habu, so
its not just new stuff that is the problem). Olympians are being ignored
for the moment, but I'll probably need to spend some time with them
sooner or later to be fair.