Monsters: Major Rebalance (BEGIN)
And so it begins. I'm tired of the only way to make the game more
difficult being to up the damage per round. Or, to up the summoning
which, by the way, leads indirectly to the first method. Can we not make
the game more difficult in an interesting way? Also, I'm damned tired of
the cryptic "monster critical hits" code for stunning and cut effects.
For example, change "B:HIT:HURT:50d1" to "B:HIT:LOSE_CHR:50d1" and
suddenly that hit insta-kills the player. No saves. No messages. No
kidding (or at least it used to until I redid the knockout code). I'll
let you puzzle out how or why.

So the focus here will be on monster melee. We'll support multiple
effects per blow and see where that leads us. Also, I'll probably redo
monster auras along the same lines. Imagine hitting a monster and having
gravity warp around you! Not to mention time auras or even curse auras.
Aura damage should be set on the monster level in r_info.txt.

Player effects will probably get a rework to. Most effects are open and
shut (think Free Action and the old fear stuff ... The new fear stuff
sets the benchmark for interesting game play changes that make the game
harder, but not in a stupid way).

OK, enough ranting. This will take a while ...