Monster Melee: Split out Confusion et.al Damage
When a titan does B:HIT:CONFUSE(12d12), he isn't exactly tickling you.
This is definitely a physical attack from a hard hitting monster and
should be treated as such: B:HIT:HURT(12d12):CONFUSE. There are a lot of
these sorts of changes ... I doubt I'm done making them.

Nerf The Unicorn of Order and DEPRECATE both the lazy personality and
The Plain Gold Ring. The Unicorn needs to be worth attempting, and she's
still very nasty, but has a lot less hp. Should be beatable. The PGR
never was and I am either removing or nerfing stupid monsters.