MON_STUNNED: Rework
Previously if a monster was stunned, it amazingly
[1] Missed 50% of its moves
[2] Failed 50% of its spells
[3] Missed 50% of its blows

Change this to:
[1] Misses 0% of its moves
[2] +50%*S/100 to fail rates (S is stun level 1 to 100)
[3] M -= M*S/150 where M is melee skill
[4] D -= D*S/150 where D is melee damage,
just like it works for the player.

Play around with monster stun recovery and monk stunning. Both of these
are liable to need a bit of tweaking. display_percentages now shows the
enemies stun pct in the health tracker.