Monster Rebalance: Compute Nastiness Index
Basically, this is the expected damage per round for the current
character, making some naive assumptions (such as monsters always use
attack spells). Currently, it ignores the effects of auras, which is
stupid (but AURA_REVENGE is tricky to handle and I don't have easy
access to the player's melee accuracy to estimate hits per round).

Still, pretty informative, and you can see the effects of AC and speed
changes. The first nastiness number is a melee strategy where the player
takes spells on freq_spell% turns, and melee on (100-freq_spell)% turns.
The second nastiness number is for a blaster mage, who takes spells at a
guesstimated (freq_spell + 12)% turns. See other spoiler docs for this
number, which is accurate for the 30% ish freq_spell range. Of course,
not all spells are offense, and monsters do other things. Also, Hand of
Doom does assumes saving throw failure ... I'll need to improve that.

This is spoiler ^A"D, btw. Spell info is derived from ^A"d. There are
two sections in the ^A"D report. The first sorts by monster level, the
second by melee nastiness (descending).