Monster Spells: Begin Total Rewrite
I tried to restrain myself on this one ... Change the monster spell
system from a flag based system to a more informative one. Spells are
grouped by type to make the spell picking AI less cumbersome. This also
facilitates mon_display(). Lore will no longer be all or nothing, but a
count of spell sightings, allowing for the option in the future of
revealing damage numbers to the player.

Spells may be parameterized in r_info. There are default damage calcs as
before, but now we can override these defaults as desired. For example,
BO_NETHER(10d10+50). The meaning should be obvious.

You can pretty much do S_<type> for any known SUMMON_<type>, just like
with the room templates. Parameterize to give the number of summons as
in S_CYBER(3d5).

For each GF_<type>, you may BO_<type>, BA_<type> or BR_<type>. Later, I
guess we'll allow BE_<type> for beams or some such. Currently, each of
these requires default damage numbers in code, but we can relax this if
needed to make adding new types easier.

This is not a stable commit, but the number of changes was growing too
large for comfort (and my editor just irreversibly deleted a large block
of code ... sigh). Code compiles. Game loads. But many spells not yet
coded and the AI still needs to be written. More to follow. And this
commit will, I guess, force the next version to a savefile breaking 7.0.