Monster Spell AI: Smart Flags
Initial implementation. Monsters flag that they know the current player
resistance, which is a bit of a cheat should that resistance change. Be
careful with SM_* constants as they changed from flags to bit positions.
And have_flag et. al. aren't quite appropriate unless you want to:
  have_flag(&mon->smart, <which>)
since mon->smart is not a u32b array.