project_p: nerf distance damage reduction (PROPOSED)
And it's not just about passwall ... If you've ever fought dragons in the
wilderness mountains then you know that the 50% damage reduction is
simply to much! Using Forest Creation is similarly beneficial in the
dungeon. I can understand 50% if the player really is incoporeal and
inside a wall, but I don't agree with this if he is hiding behind a tree
or even a spider web ... Still, passwall is too strong, so even if the
50% number is realistic, it's getting a nerf ;)