Possessor: Begin Long Rewrite to New Monster Stuff
In the process, I decided to remove both the Imitator and the Blue-Mage
from the game. I never enjoyed the former class (play a mimic instead!)
and the latter was OK, but it will require a total rewrite to bring it
back. If somebody asks, I might ...

The new possessor will more closely follow r_info for melee, auras and
spells. Wrt melee, even the innate_attack system I did 5 years ago is
not up to the new monster melee system. Plus, things just got too
complicated in innate_attacks for my tastes.

Spells are going to be a challenge, requiring a bit of UI work. This
will take a while ...