Monsters: Replace DR_MANA with DRAIN_MANA
Note, the spell, DRAIN_MANA works just like the old melee effect,
DR_MANA, except that it also heals the caster. So a side effect of
this change is that the melee effect now also heals the attacker.
Consistency is probably a good thing ... If not, we can add an
if (touch) guard in gf_damage_p.

I'm trying to simplify game mechanics ...