Possessor: Custom Spells
Things are easier if I make possessor spells a new spell group with
a default monster allocation probability of 0. For one thing, some
of these spells could become monster spells (e.g. Berserk). For another,
this vastly simplifies the spellcasting UI. Finally, it allows us to
parameterize spells in the normal way (This is not currently needed,
but who knows?). So massive changes to r_info moving F: tags to S: tags.