MTIMED: Cleanup monster timers
I don't think the mproc stuff is a good optimization, and have no
evidence that it is still needed. Better if we just process monster
timers in process_monsters() since we are looping thru them all
anyway. The old optimization wasted alot of space as well.

I'm preparing for new timed effects.