Monsters: Clarify r_idx vs ap_r_idx
It was unclear to me when to use which. Makes for some awkward coding
as well, especially with the oversized functions (i.e. variables at the
top might be declared 1000+ lines before they are used, and you need to
inspect the usage to decide which field is correct). Details in code.

Some playtesting as well ... Problems with monster pain messages.