Monster Spell AI: Uncover the Player
One way to defeat the Metal Babble is to use forest creation and
simply wait for the babble to step next to the player for a needle
poke. Of course, the babble has a Fire Ball spell, and could easily
defeat this strategy if it were 'SMART' enough. Now, alas, it is!

Monsters with the SMART flag will splash appropriate balls and breaths
to uncover players hiding out of line of sight behind either trees or
spider webs. Not every possible gf is supported since some only work
on trees and not webs, or vice versa. But enough are to make this a
very evil AI change.