Monster Anger: Adjust for early and mid-game
The old progression had the end game in mind, but this progression was a
bit harsh on weak spellcasters. For example, a mindcrafter only does
about 70 damage per round until the late CL30's. Taking down 1000hp
breathers ends up making 10% breathers into 70% or even 80%, especially
if the player is fast (and mindcrafters usually are). You'll need a
spreadsheet to evaluate this change.