Monsters: Remove summon_ct
This was an old attempt to limit monster summoning, but it never
actually worked very well. These days, the spell probabilities and
number of respondees have been tweaked for a better solution.

If you are curious, the problem was that summoners summon summoner
summoning summoners, which, in turn summon more summoner summoning
summoners. So that the exponential blow up of summoned monsters never
really got counted correctly (i.e. summons should have been rooted
in a tree starting with the original summoner, in which case the limit
approach may have worked OK).