check_experience() ... continued
There are 174 calls to handle_stuff in the system. Turns out that it
still gets called all the time during melee attacks. What the code wants
is to "redraw_stuff" and "window_stuff" ... very seldom to
"update_stuff" (although some rare classes want this as well). Let's
move PU_EXP to PN_EXP and check_experience() in notice_stuff instead.
This is only called 14 times in the system, and seemds a better home for
this change.