Archery: Make Crossbows less Awesome
First of all, archery is a bit OP despite what other players claim.
However, this OP is most manifest wrt crossbows and bolts, since the
ammo quality (especially on 7d5 bolts) blows up with slays and crits.

For this commit, change the calculation from:
   M*(XdY*S*C + D1) + D2
to
   M*(XdY + D1)*S*C + D2
In general, this requires roughly halving(1) the S and C multipliers, but
it makes slings (5d3+15) much more compelling than they were, since
slays and crits work on 25 mean damage rather than 10.

(1) Halving means halving the bonus: S' = 1.0+(S-1.0)/2. So 2.0x slay
evil is now 1.5x. 3.0x slay dragon is now 2.0x. Etc.

(2) This is a bit of retrenchment from my "make ammo dice matter"
initial design. But, actually, my complaint is to make "ammo matter",
and we still have that. In the old days, 50-60% of archery damage came
from the +to_d on the *shooter*, so you didn't care much about ammo
(unless you got a slay, which boosted that 60% as well, but a lack of
quivers made slay ammo not usable).