Archery: Continued
Next, one of the complaints was that the gap between good ammo and
excellent ammo is too large. I agree with this one ... Oftentimes a slay
or brand nearly doubles your damage output, and a Kill Foo is ridiculous
(e.g. 274% increase).

To address this, we shall move some power from S into C:
  M*(XdY + D1)*S*C + D2
In the case S is 1.0 (no slay) this will increase the baseline damage
for normal ammo. Decreasing S so that S * C is approximately the same
will keep the same overall damage for excellent ammo.

Example: Elemental brand is now a 44% damage boost (was 83%). Of course,
the exact percentages will vary ... I'm taking some test characters.