Document: Fix drawing bug with no wrap lines
You'll see this problem in inventory/equipment listings, which don't
support word wrapping. The last character on the line would continually
be over-written by the first character in subsequent, off-the-edge word
tokens. Token processing must continue to properly handle style
directives, but, once we have run off the edge, we should skip any
subsequent word tokens until an explicit newline is received. This is
only a problem for the NO_WRAP style.