Monster Spells: Adapt Frequency after Repeat Casting (EXPERIMENTAL)
Simulate mana by keeping a counter. As the monster casts, the counter
increments, and the effective spell frequency goes down. Fail to cast,
and the counter decrements, and the effective spell frequency begins to
return to the baseline. Anger a monster, and the counter is ignored.

Motivation? 50% spell casters often go on a tear, casting many many
times in a row. I'm attempting to cool this down a bit. To be fair, the
frequency should probably rise if the monster fails to cast too often
... but we have mon->anger to address this (and melee guys seem to need
a break these days).

This will need playtesting to evaluate.