Monster Lore and Spell Splashing
We lore the splashes, but not those turns the monster failed to splash.
The result is grossly overstating the spell frequency for certain
monsters, like Sorcerers or Dark elven shades (e.g., reporting 80% when
the actual observed frequency is 30%). This is manifest if you use ASC
to control the summons.

Another lore problem happens when monsters battle each other. The spells
are lored, but moves are not (e.g. Pip vs a pack of vibration hounds and
suddenly the hounds breathed 81 out of 81 moves!). Now, we only lore mon
vs. mon spells if the player is projectable to the caster. Of course, we
always lore mon vs plr spells, even if the player is not projectable,
since we want some record of all the splash casting going on.

Seems to work better, but needs extensive playtesting.

Finally, improve wizard monster spell ai. If the monster can't cast,
don't prompt the wizard (that's me). But if they can, then report the
spell the default monster ai wanted to use (in violet).