Object Allocation and Monster Races (EXPERIMENTAL)
Object allocation first chooses an object kind using an allocation table
(so many potions, so many weapons, etc.). For many monster forms, the game
is a bit frustrating, especially if you can not wield 90% of the equipment
you find. To address this, let's dynamically adjust the allocation
probability down for unusable classes of objects. This might require
tweaking, but initial data looks promising.