Auto-mapping: Give option to prevent auto usage
In the early game, one may carry a small pile of ?Mapping for strategic
usage in dangerous situations. While auto-mapping is very convenient
most of the time, it would be nice to be able to turn it off (could be
another option, I guess). For now, simply inscribe '$' on the object to
prevent its usage. (This only works for mapping, not for auto-id).