Monster Lore and ASC AI
Problems loring monster moves in ASC. Previously, we would lore spells,
but not moves when the monster is not projectable. This would inflate
the observed spell frequency if the player rested for monsters to
approach since, eventually, the monster will splash the player. However,
the fix to this was over-reporting the moves. A monster will not move
from a non-player-projectable grid into a player-projectable grid unless
there are enough open squares (and the monster can summon). But this
scenario does not mean that monster has a valid splash opportunity
against the player (e.g. ASC with a 'knights move' projection
opportunity at the tail). So we need to check splash-ability as well.
Finally, the location of the ASC AI was problematic wrt passwall, I
think. At any rate, something weird was going on that I still don't
understand, so I moved the AI code into the 'can_cross' block of the
larger move AI. Brief playtesting seems to work, but I'll need to keep
an eye on this one.