reduce_uniques: also reduce number of bounty uniques
We'll keep the tail of the list, so the player can get the best prizes.
But if the player reduces uniques to N% of the total, then the bounty
list is similarly reduced. For example, 50% uniques has a bounty list of
just 10 wanted uniques. 10% uniques has only 2, but the player can still
get the ?ArtifactCreation.