High Scores: What if user creates html dump for oook?
In this case, the html attributes that oook relies on for posting will
be missing, since doc formats are just lightly salted text files.
However, the scores.txt database does (now) have enough info to
reconstitute all the attributes that I think oook needs. The only doubt
is the status attribute (which replaces the 3 winner, dead, retired
boolean attributes).

The scenario is that the user died and missed the dump (how, I have no
idea). In this case, they can recover the death dump from the scores
directory, but it is in doc format not html. But view the file in ~H and
then pipe to html (|), and the user can now post to oook ... hopefully.