Mimics: Browse available forms
Mimics need a way to browse available forms (known or potential) without
first stopping their current mimicry. The old way was tedious, and
bloats the history portion of the character dump when all the player is
trying to do is compare their known forms. I added a new power to browse
forms that is only available when mimicry is active. It uses no energy.