Devices: Pricing
Work a bit on pricing. With devices available for longer level ranges,
the charge calculation was blowing up the values of certain items (e.g.
Rod of Fireballs 180dam for 50k). Some of this was also rounding issues
for rods, so scale charges by 10. As always, object scoring is a work in
progress ...