"Magi: Magic Mastery pvals decreased
In the old days, Magic Mastery was available two ways:
[1] Amulet of Magic Mastery
[2] Gloves of Magic Mastery
In both cases, the objects did *nothing else*, so you effectively
had to lose a slot to boost your device skills. These days, we have
gloves of the wizard which boost devices skills and int, but also
decrease combat bonuses and perhaps str and con as well. That's good
balance IMO, since warriors will think twice on that one. But "Magi is
different, granting OF_MAGIC_MASTERY with no downside. Often, you can
get something awesome like telepathy in the bargain.

I've gotten too many +5 mastery amulets. Let's reserve this elite status
to "Zeus and limit "Magi to +4 max, which is now as rare as +5 used to
be.