Shopping: Black Market and Jeweler Object Levels
This is small code change with potentially big effects (might be broken
in the end game, so we'll need to keep an eye on this one).

BM and Jeweler stocking now uses the player level to compute object
levels to stock. I saw this idea in PosChengband R, but the code used
there was too generous. Here is the problem as I see it:

[1] The BM stocked at 25+1d25 ... always. This rewarded players in the
    early game. After all, where can a mid level player get an L50
    object? Certainly not on DL40. Even on DL50, they are hard to come
    by. OL50 is normal around DL56, and that assumes you can kill L45+
    monsters to get them (or can raid vaults).

[2] My playing experience is to overshop in the early game. Pre-Smaug,
    a good percentage of my gear is town bought. I'm a bit cautious,
    doing stat gain on DL39 of the Lonely Mountain (which is still very
    challenging at CL35, but levels are completely clearable most of the
    time). Getting L50 objects in town easily beats what I get in the
    dungeon. And with 5 towns with 5 black markets and 1 jeweler, I do
    alot of travelling. Bleck!

[3] Late game, the opposite obtains. The BM and J become useless, except
    for keeping me well stocked with !Healing, ,Restoring, et. al.

What we need is to nerf the early game stocking a bit, and reward late
game stocking. You can now get L69 objects in the BM ... at CL50. But
you need to be CL39 before L50 objects can spawn, and CL20 characters
can get OL29 at best. I'm sure this will be labelled a 'nerf' by most,
but the end game is actually a generous 'buff'!