Artemis: No attacks? Really? Well, then don't AI_SEEK!
I'm not sure if this was an oversight by Dave or myself, but as of the
initial PosChengband load, Artemis has had no melee attacks. If she is
allowed to AI_SEEK, then she will never shoot if the player uses an ASC.
When she gets close, she will either blink away, or engage with her null
attacks for 0 hits and 0 damage. Quite easy this way, actually.

Force AI to be one of AI_SHOOT, AI_LURE or AI_MAINTAIN_DISTANCE. I guess
we could give her some blows, but I think this will be more interesting.