Devices: More Mercy
Or not ... Device skills are indeed more important than they were in
Hengband. So I think the racial and personality bonuses should be
adjusted. After all, -10 devices is way more meaningful than +20 melee
skill (which is about +7 to hit). That's how Hengband did the Mighty
personality. I've made it -8 devices for +30 melee, and still, that is a
hard choice to make. Shrewd is still +16 devices as the 97% life rating
is harsh, especially for mage types. Races generally get 70% of the
bonus (or penalty) as before. That's the not.

As for the mercy, I'll allow the device difficulty calculation to be
parameterized with a cutoff value. Set to 0 for the old version. This
gives noticeable effects at around _Healing difficulty (generally 8%
higher fails). Set to 40 to allow entry level healing as before, while
still making the tail of the device curve very challenging.

As usual, see design/devices.ods or, better yet, ^A"2 for online tables.