Devices: Adjust INT bonuses
I always felt INT was less important for devices than it should be.
I'm also puzzled as to why an INT of 8 gives a bonus to devices ...
that's pretty stupid. And INT of 3 gives no penalty?

Slight adjustments to the INT device table, making 18/*** worth 23
rather than 20, and 3 worth -5 rather than 0. Most of the table is
unchanged.