Devices: Fix absurd charge counts on high power devices
For example, a rod of beam of light with 12 charges for 250 damage. The
problem is that many effects have a large window of availability, and
strongly scale damage with device power. For example, a rod of beam of
light for 35 damage (cost = 15). 15sp for 35 damage is OK. 15sp for 250
damage is not ... I did a quick solution for now, but I should probably
think more about this one.