Nexus Scramble: Scramble Location rather than Stats
Nexus hounds are DL27, well before stat gain. Scrambling stats has been
called "the worst game mechanic EVER" and I have to agree with that in
the sense that it totally ruins the game most of the time. At least when
it happens early. You can survive later game scrambles, especially once
stat gain is well underway. I think Vanilla made the effect temporary,
but I don't like that approach since player's can just wait it out.

Let's try the following:
[1] Nexus scrambles the player's location with teleport, teleport to and
teleport level. We'll add teleport to new dungeon as well. Of course,
this is only possible if wilderness is enabled. (no_wilderness is a
non-standard option and will keep mutate_player).
[2] Pick a different dungeon type, excluding Angband (due to quest
issues), random dungeons, and winner dungeons (heaven and hell).
[3] Pick a new level in the valid range for this dungeon type. There is
some mercy wrt a max level to place the player, but you can go out of
depth.
[4] Player teleports to this dungeon and level instantly.
[5] If the new dungeon has never been visited before, and has a still
living surface guardian, then Word of Recall will be blocked. Player
must travel back to the surface manually and walk back to nearest town.
[6] Player does gain the new recall point, so even if they don't kill
the guardian, they'll be able to return to the location of the nexus
travel. However, again, Word of Recall will be blocked until the surface
guardian is slain.

This effect can be both annoying and deadly. Hopefully, it is not "the
worst game mechanic EVER" :)